Rename sign_der() to ToDer::signed()
